Updated so non-logged in users can use site - won't redirect home to login page, will redirect comment and like to login page.

+++ website/views.py
@@ -8,7 +8,6 @@ views = Blueprint("views", __name__)
 
 @views.route("/")
 @views.route("/home")
-@login_required
 def home():
     posts = Post.query.all()
 
@@ -68,7 +67,8 @@ def create_comment(post_id):
     text = request.form.get("text")
 
     if not text:
-        return jsonify({"error": "Comment cannot be empty!"}, 400)
+        flash("Comment cannot be empty!", category="error")
+        # return jsonify({"error": "Comment cannot be empty!", "status": 400})
     else:
         post = Post.query.filter_by(id=post_id)
         if post:
@@ -76,9 +76,11 @@ def create_comment(post_id):
             db.session.add(comment)
             db.session.commit()
         else:
-            return jsonify({"error": "Post does not exist!"}, 400)
+            flash("Post does not exist!", category="error")
+            # return jsonify({"error": "Post does not exist!", "status": 400})
 
-    return jsonify({"text": text, "comment_id": comment.id})
+    return redirect(url_for("views.home"))
+    # return jsonify({"text": text, "comment_id": comment.id, "status": 200})
 
 
 @views.route("/delete-comment/<comment_id>")
@@ -87,25 +89,33 @@ def delete_comment(comment_id):
     comment = Comment.query.filter_by(id=comment_id).first()
 
     if not comment:
-        return jsonify({"error": "Comment does not exist."}, 400)
+        return jsonify({"error": "Comment does not exist.", "status": 400})
     elif current_user.id != comment.author and current_user.id != comment.post.author:
-        return jsonify({"error": "You do not have permission to delete this comment."}, 400)
+        return jsonify({"error": "You do not have permission to delete this comment.", "status": 400})
     else:
         db.session.delete(comment)
         db.session.commit()
 
-    return jsonify({"success": "Comment deleted."})
+    return jsonify({"success": "Comment deleted.", "status": 200})
 
 
 @views.route("/like-post/<post_id>", methods=["POST"])
-@login_required
 def like(post_id):
+    # Initially check that user is logged in. Return error if not.
+    if not current_user.is_authenticated:
+        flash("You must log in to like posts!", category="error")
+        return jsonify({"error": "You must log in to like posts!", "status": 401})
+
+    # Retreive the post and like objects given the post id and current user's id
     post = Post.query.filter_by(id=post_id).first()
     like = Like.query.filter_by(author=current_user.id, post_id=post_id).first()
 
+    # Check that the posts exists. Return error if not.
     if not post:
-        return jsonify({"error": "Post does not exist."}, 400)
-    elif like:
+        return jsonify({"error": "Post does not exist.", "status": 400})
+
+    # Update the like count.
+    if like:
         db.session.delete(like)
         db.session.commit()
     else:
@@ -113,4 +123,7 @@ def like(post_id):
         db.session.add(like)
         db.session.commit()
 
-    return jsonify({"likes": len(post.likes), "liked": current_user.id in map(lambda x: x.author, post.likes)})
+    # Return the amount of likes and bool of whether user has like the post or not.
+    return jsonify(
+        {"likes": len(post.likes), "liked": current_user.id in map(lambda x: x.author, post.likes), "status": 200}
+    )
